put set cookie code into separate file
